Add files via upload
At the current moment, I've completed the framing for the mobile site, from here I'll be creating media queries for both the tablet and desktop versions. I think getting that to work before I add content in would be a lot smarter than content first and then converting that to other views. I had to remove my js folder with main.js in it since I had nothing in there currently and so it wouldn't let my upload a blank file.